rcar_gen3: Remove __user macro from the variable declaration
__user macro is necessary to refer and update data of address destination.
Since cb_func and user_data defined in the vspmif_drv driver do not refer
to the contents of the address, the __user is unnecessary in the variable
declaration. Also, although vspm_if.h is included in the user layer,
since __user is not defined in the user layer, a compile error occurs.
Therefore, __user macro should be deleted from the variable declaration.

Signed-off-by: Sojiro Kusunoki <sojiro.kusunoki.zj@ps.hitachi-solutions.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>